feat: adding global logger
adding winston global logger in a more simplified version
