+++ src/middleware/logger.ts
@@ -1,11 +1,11 @@
-import logger from '@/utils/logger/logger';
-import { NextRequest, NextResponse } from 'next/server';
+import logger from '@/lib/winston/global-logger';
+import type { NextRequest, NextResponse } from 'next/server';
 
-export function logRequest(
-  req: NextRequest,
-  res: NextResponse,
-  next: () => void
-) {
-  logger.info(`Incoming request: ${req.method} ${req.url}`);
-  next();
+export function middleware(req: NextRequest, res: NextResponse) {
+  logger.info('Incoming request', {
+    method: req.method,
+    url: req.url,
+    headers: req.headers,
+  });
+  return res;
 }
